code/src/requirements.txt
flask==3.0.0
pywin32==310
pandas
extract-msg==0.41.1
beautifulsoup4==4.12.2
mail-parser==4.1.2
pdf2image==1.16.3
Pillow
pytesseract==0.3.10
PyMuPDF == 1.18.19
pdfplumber==0.5.28
python-docx
google-generativeai
grpcio
openpyxl
